Updated source
This commit moves the audio handling from pydub to pygame. It also updates the UI to be more pixel perfect, and added an option to enable / disable fading. Bunch of bugs fixed, and added arguments. This commit also adds a requirements.txt which will be needed later if people want to use this thing